+++ tests/extra_python_package/test_files.py
@@ -34,6 +34,7 @@ main_headers = {
     "include/pybind11/pytypes.h",
     "include/pybind11/stl.h",
     "include/pybind11/stl_bind.h",
+    "include/pybind11/stl/filesystem.h",
 }
 
 detail_headers = {
